pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.2",
    "langchain-classic>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.24",
]
